requirements.txt
asgiref==3.3.4
attrs==21.2.0
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.12.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.2
djangorestframework==3.12.4
drf-spectacular==0.16.0
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.0
jsonschema==3.2.0
lxml==4.6.3
MarkupSafe==2.0.0
pyrsistent==0.17.3
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.4
